+++ ndsl/dsl/dace/__init__.py
@@ -1,4 +1,9 @@
-from .dace_config import DaceConfig
+from .dace_config import DaceConfig, DaCeOrchestration
 from .orchestration import orchestrate, orchestrate_function
 
-__all__ = ["DaceConfig", "orchestrate", "orchestrate_function"]
+__all__ = [
+    "DaCeOrchestration",
+    "DaceConfig",
+    "orchestrate",
+    "orchestrate_function",
+]
