+++ ndsl/dsl/dace/utils.py
@@ -11,7 +11,7 @@ from gt4py.cartesian.gtscript import PARALLEL, computation, interval
 import ndsl.xumpy as xp
 from ndsl import ndsl_log
 from ndsl.config import Backend
-from ndsl.dsl.dace.dace_config import DaceConfig
+from ndsl.dsl.dace.dace_config import DaceConfig, DaCeOrchestration
 from ndsl.dsl.stencil import CompilationConfig, FrozenStencil, StencilConfig
 from ndsl.dsl.typing import Float, FloatField
 
@@ -19,13 +19,13 @@ from ndsl.dsl.typing import Float, FloatField
 class DaCeProgress:
     """Rough timer & log for major operations of DaCe build stack."""
 
-    def __init__(self, config: DaceConfig, label: str) -> None:
-        self.prefix = DaCeProgress.default_prefix(config)
+    def __init__(self, mode: DaCeOrchestration, label: str) -> None:
+        self.prefix = DaCeProgress.default_prefix(mode)
         self.label = label
 
     @classmethod
-    def default_prefix(cls, config: DaceConfig) -> str:
-        return f"[{config.get_orchestrate()}]"
+    def default_prefix(cls, mode: DaCeOrchestration) -> str:
+        return f"[{mode}]"
 
     def __enter__(self) -> None:
         ndsl_log.debug(f"{self.prefix} {self.label}...")
@@ -40,9 +40,12 @@ def _is_ref(sd: dace.sdfg.SDFG, aname: str) -> bool:
     for node, state in sd.all_nodes_recursive():
         if not isinstance(state, dace.sdfg.SDFGState):
             continue
-        if state.parent is sd:
-            if isinstance(node, dace.nodes.AccessNode) and aname == node.data:
-                return True
+        if (
+            state.parent is sd
+            and isinstance(node, dace.nodes.AccessNode)
+            and aname == node.data
+        ):
+            return True
 
     return False
